Add failing tests to mark what to implement with security
I need to restrict viewing meetings and members to the unit associated
with the logged-in user. These tests are just stubs right now, but
they'll be grown into tests to make sure that I'm not getting back
more than what I should.

Side note: there's some stuff on the stash that I'll probably want to
pop and work off of.